Refactor project running common handlers
